move bayes net test out of inference test
